+++ fiscalsim_us/variables/gov/states/va/tax/income/va_exemptions.py
@@ -24,7 +24,7 @@ class va_exemptions(Variable):
         spouse_if_filing_jointly = where(
             (filing_status == filing_statuses.SINGLE)
             | (filing_status == filing_statuses.HEAD_OF_HOUSEHOLD)
-            | (filing_status == filing_statuses.WIDOW)
+            | (filing_status == filing_statuses.SURVIVING_SPOUSE)
             | (filing_status == filing_statuses.SEPARATE),
             0,
             where(filing_status == filing_statuses.JOINT, 1, 0),