+++ fiscalsim_us/variables/household/income/household/household_state_income_tax.py
@@ -9,6 +9,7 @@ class household_state_income_tax(Variable):
     unit = USD
     definition_period = YEAR
     adds = [
+        "ar_income_tax_before_refundable_credits",
         "ca_income_tax_before_refundable_credits",
         "co_income_tax_before_refundable_credits",
         "dc_income_tax_before_refundable_credits",
@@ -35,6 +36,7 @@ class household_state_income_tax(Variable):
         "or_income_tax_before_refundable_credits",
         "pa_income_tax",
         "ri_income_tax_before_refundable_credits",
+        "sc_income_tax_before_refundable_credits",
         "ut_income_tax_before_refundable_credits",
         "va_income_tax_before_refundable_credits",
         "vt_income_tax_before_refundable_credits",
@@ -43,39 +45,40 @@ class household_state_income_tax(Variable):
         "nyc_income_tax_before_refundable_credits",
     ]
     subtracts = [
-        "ca_refundable_credits",  # California.
+        "ca_refundable_credits",  # California
         "co_refundable_credits",  # Colorado
-        "dc_refundable_credits",  # District of Columbia.
-        "ia_refundable_credits",  # Iowa.
-        "il_refundable_credits",  # Illinois.
-        "in_refundable_credits",  # Indiana.
-        "ks_refundable_credits",  # Kansas.
-        "ky_refundable_credits",  # Kentucky.
-        "la_refundable_credits",  # Louisiana.
-        "ma_refundable_credits",  # Massachusetts.
-        "me_refundable_credits",  # Maine.
-        "md_refundable_credits",  # Maryland.
-        "mn_refundable_credits",  # Minnesota.
-        "mt_refundable_credits",  # Montana.
-        "mo_refundable_credits",  # Missouri.
-        # Skip NC, which has no refundable credits.
-        "nd_refundable_credits",  # North Dakota.
-        "ne_refundable_credits",  # Nebraska.
-        "nh_refundable_credits",  # New Hampshire.
-        "nj_refundable_credits",  # New Jersey.
-        "nm_refundable_credits",  # New Mexico.
-        "ny_refundable_credits",  # New York.
-        "ok_refundable_credits",  # Oklahoma.
-        "or_refundable_credits",  # Oregon.
-        # Skip PA, which has no refundable credits.
-        "ri_refundable_credits",  # Rhode Island.
-        "ut_refundable_credits",  # Utah.
-        "va_refundable_credits",  # Virginia.
-        "vt_refundable_credits",  # Vermont.
-        "wa_refundable_credits",  # Washington.
-        "wi_refundable_credits",  # Wisconsin.
+        "dc_refundable_credits",  # District of Columbia
+        "ia_refundable_credits",  # Iowa
+        "il_refundable_credits",  # Illinois
+        "in_refundable_credits",  # Indiana
+        "ks_refundable_credits",  # Kansas
+        "ky_refundable_credits",  # Kentucky
+        "la_refundable_credits",  # Louisiana
+        "ma_refundable_credits",  # Massachusetts
+        "me_refundable_credits",  # Maine
+        "md_refundable_credits",  # Maryland
+        "mn_refundable_credits",  # Minnesota
+        "mt_refundable_credits",  # Montana
+        "mo_refundable_credits",  # Missouri
+        # Skip NC, which has no refundable credits
+        "nd_refundable_credits",  # North Dakota
+        "ne_refundable_credits",  # Nebraska
+        "nh_refundable_credits",  # New Hampshire
+        "nj_refundable_credits",  # New Jersey
+        "nm_refundable_credits",  # New Mexico
+        "ny_refundable_credits",  # New York
+        "ok_refundable_credits",  # Oklahoma
+        "or_refundable_credits",  # Oregon
+        # Skip PA, which has no refundable credits
+        "ri_refundable_credits",  # Rhode Island
+        "sc_refundable_credits",  # South Carolina
+        "ut_refundable_credits",  # Utah
+        "va_refundable_credits",  # Virginia
+        "vt_refundable_credits",  # Vermont
+        "wa_refundable_credits",  # Washington
+        "wi_refundable_credits",  # Wisconsin
         # LOCAL
-        "nyc_refundable_credits",  # New York City.
+        "nyc_refundable_credits",  # New York City
     ]
 
     def formula(tax_unit, period, parameters):