+++ fiscalsim_us/variables/household/income/household/household_tax_before_refundable_credits.py
@@ -12,6 +12,7 @@ class household_tax_before_refundable_credits(Variable):
         "employee_payroll_tax",
         "self_employment_tax",
         "income_tax_before_refundable_credits",  # Federal.
+        "ar_income_tax_before_refundable_credits",
         "ca_income_tax_before_refundable_credits",
         "co_income_tax_before_refundable_credits",
         "dc_income_tax_before_refundable_credits",
@@ -38,6 +39,7 @@ class household_tax_before_refundable_credits(Variable):
         "or_income_tax_before_refundable_credits",
         "pa_income_tax",  # PA has no refundable credits.
         "ri_income_tax_before_refundable_credits",
+        "sc_income_tax_before_refundable_credits",
         "ut_income_tax_before_refundable_credits",
         "va_income_tax_before_refundable_credits",
         "vt_income_tax_before_refundable_credits",