+++ fiscalsim_us/variables/household/income/person/retirement/keogh_distributions.py
@@ -1,0 +1,9 @@
+from fiscalsim_us.model_api import *
+
+
+class keogh_distributions(Variable):
+    value_type = float
+    entity = Person
+    label = "Keogh plan distributions"
+    unit = USD
+    definition_period = YEAR